requirements.txt
flask>=2.3.0
psycopg2-binary>=2.9.7
gunicorn>=21.2.0
requests>=2.31.0
schedule>=1.2.0
